Add --post option to allow postprocessing from CLI
This version of pss_console.py has been edited to allow for postprocessing from the command line.
The program runs as originally designed unless the additional --post argument is added to the call.
--post causes the program to look for .PlanetarySystemStacker.ini in the users home directory
if the .ini file does not exist the program exits with a warning that a existing .ini is required
if .ini is found program loads those configurations and allows for changes to be made to the
    configurations via arguments (stacking only) as in the original structure
The end result is automated postprocessing and save_postprocessed_image with previously saved
    configurations settings when --post is added as an argument

Chris Mandrell, Southern Illinois University-Carbondale
School of Physics and Applied Physics
8/23/2023